Guard against buffer writes out of range.
=== Files changed ===

	modified:   linux/xterm.c

1. Modified plcchr() to check buffer size before placing character.

=== Bug numbers affected ===

=== Notes ===

The buffer size is (presently) fixed in xterm.c. The buffer sizes are
set to 250x250, so that handles most screen sizes. Having a buffer size
less than the onscreen size is not fatal, it will just treat the
displayed text as a rectangle within the window (a mode that needs to
be checked!). However, using virtual space mode, there are lots of
reasons we could be out of bounds, perhaps by a lot.