Fix randn() overflow; turn off cursor; lower rate.
=== Files changed ===

	modified:   terminal_programs/wator.c

1. Fixed bug in randn(). It works on 32 bit, but in 64 bit, the function
gave results larger, and the result is the limit calculation overflowed.
The fix was to carry out the calculation in long, which is 64 bits on a
64 bit machine (but see note 1 below).

2. Turned off the cursor during the run, so you don't get flashing
effect. This is pretty standard, actually.

3. Turned the rate down to 10 times per second, which is enough. It
handled the 30 times/sec rate, but it was actually to fast to reasonably
watch, and, moreover, it shut out the ctl-c key, which you need to
get out of this demo. This may argue for rate limitation!

=== Bug numbers affected ===

=== Notes ===

1. This calculation is suspect, since long is 32 bits on a 32 bit
machine. Mingw didn't overflow because RAND_MAX was low enough, but
it could have been higher. I suspect widening to long long is the
answer, since this is carried out as 64 bit even on a 32 bit machine.

2. Wator is actually the first program I ever ran on the terminal
handler in 1984! In fact, I created it for that purpose. See online
comments.