Add programs to makefile all; clear screen; flip screen; fix for console.
 Changes to be committed:
	modified:   Makefile

1. Added wator to build all.

	modified:   terminal_games/pong.c

1. Clear screen.

	modified:   terminal_programs/wator.c

1. Flip screen. Fixes issues. See next.

	modified:   windows/console.c

1. Calculate what part of buffer is in display, then only write to
that. Fixes issues with consoles with active scrollback buffers.
See notes.

Notes:

I put in a fix for the problems with console windows with scrollback
buffers. It works, but only on screens where you flip screens, and
frankly the results are irrational. Windows seems to adjust the buffering
according to the area you are writing to, but even this makes no sense,
because the display offset works even though the buffer has changed.

Obviously this needs further research, but the fix works for now.