Fix debug issue; remove debug from scntst; fix console issues.
1. Include our stio.h so that modules can be debugged.

1. Removed debug stops from test.

1. Removed getpos and saves for cursor position from that. See notes.

2. Tried to stop console wrapping, not successful. Have to do it manually.

1. Removed fixed issues.

Notes:

getpos was an attempt to have console.c cooperate with other programs
setting things on the console. It didn't work. It just slowed the program
and caused things like the true size of the screen being out of sync with
our idea of it, and odd program behavior.

The new mode is: we own the console until we release it.