Updates for Mac OS X
 Changes to be committed:
       modified:   Makefile

1. Created new test for OS type, and put resulting ID in OSTYPE.
This replaces test for OS environment variable, which only works
for Windows, thus this test was only Windows or non-windows.

2. Changed OS specific builder clauses. Mac OS X generally follows
Windows, but the rather braindamanged syntax of make makes it easier
to just create null clauses.

3. Create a series of new product builds for Mac OS X. Most of these
evaluate to stub files, which is of course why the stubs directory
was created.

       modified:   libc/stdio.c

1. Fixed expression association error.

       modified:   linux/services.c

1. Got rid of '=' association warning.

2. Fixed incorrect '=' operator to '==' (good catch clang).

       modified:   stub/graph.c
       modified:   stub/network.c
       modified:   stub/services.c
       modified:   stub/sound.c
       modified:   stub/terminal.c

1. Fixed warnings about missing returns. The stubs do not need this,
error() does not return, but clang warns.

       modified:   tests/svstst.c

1. Fixed prttimdat() parameter type.

2. Fixed all internals for static. See next.

3. Changed name of wait() to waittime() because there is a sys routine
of that name. This EVEN THOUGH IT IS STATIC! This sucks big time. Apparently
the clang rules are different. There are several of these overlaps in
Petit_ami that are going to have to be changed.

Notes:

1. Not yet working. There was a proof of working principle done back when
petit-ami on Mac OS X from like 10 years ago! May take some study to
get it to work.

2. Services will likely work because Mac OS X is Posix (I believe!).
Am going to try out xterm under Mac OS X. It will probably work, but
may have trouble with "exotic" things like mouse positions.