Removed test from makefile; removed linux clauses from xterm.
=== Files changed ===

	modified:   Makefile

1. Removed debug setup from makefile.

	modified:   linux/xterm.c

1. Removed __linux__ clauses from xterm.c. See notes.

=== Bug numbers affected ===

=== Notes ===

The __linux__ #ifdef sections in xterm were to compile under windows,
which was an aborted attempt to make it run under windows, it didn't
work.

In any case, it would have had to be modified for Mac OS X, since that
is not Linux and does not have the define. So removed them.

xterm does not work under Mac OS at this time. The difference appears
to be that the timer calls are different than linux, and these calls
apparently are not Posix in any case.

The timer calls (if xterm can be ported at all) need to be (apparently)
refactored to BSD calls, which are very different. Most of the Mac OS X
calls can be referenced to BSD, since Darwin came from that direction.