Changed the way standard fonts works; added char attributes; fix justify bug.
 Changes to be committed:
	modified:   include/graph.h

1. Remove fixed standard font defines and add functions for same instead.

	modified:   tests/gratst.c

1. Fixed bad translate.

2. Remove fixed standard font defines and add functions for same instead.

3. Add eolns to extended effects tests.

	modified:   windows/graph.c

1. Added new character screen attributes.

2. Removed progressive lines system, not used in C without overloads.

3. Added internal numbers for standard fonts.

4. Translate new character weights to Windows weights. This is the only
new screen character attribute that actually does something. The others
are (hollow, raised, etc.) are filled out but do nothing.

5. Pretty much refactored the standard fonts system. Now the fonts are
loaded in a list as Windows gives them (with @fonts removed), and the
standard fonts are just calls that yeild the numbers of the fonts.
See note below.

6. Added new standard font routines.

7. Fixed bugs in iwritejust() and ijustpos().

8. Filled out the previously no-oped screen character attribute calls.

Notes:

The standard fonts system has been a sick puppy for a long time. I don't
like to change APIs, but this one needed to be done. The straw that broke
the camel's back was the need to alias the tech font with the sign font,
which even the old manual mentions. This was very difficult to fit into
the old system.

In any case, the old system made it hard to count and list fonts, ruined
any ability to alphabetically order them, and was too complex besides.
Now we just form a straightforward font list from windows, and the standard
fonts turn into functions that give those font numbers. The numbers are
found in advance, and the net overhead is mainly the locking.