Fixed fgets() bug; further compile errors config.c; bug fix config.c.
 Changes to be committed:
	modified:   Makefile

1. Moved test back to plain.

	modified:   libc/stdio.c

1. Fixed error that fgets() should not return EOF if any characters are
read. Basically, it means fgets can return a line that has no terminating
eoln, and not return EOF on that situation. This is in the man pages, and
its up to the programs to compensate.

	modified:   utils/config.c

1. Removed subcomments from header.

2. Add includes.

3. Move config tree description to config.h.

4. Various fixes to parlab.

5. Create addend() routine and move from pushing to front of list
to adding to back of list. This means that config items will appear
in the same order as in the file, which is easier to understand.

6. Refactor parlst().

7. Various other refactor.