Add runtime dialog errors; translate manlib.c
 Changes to be committed:
	modified:   petit_ami.cfg

1. Added default setting for dialog runtime errors.

	modified:   tests/mantst.c

1. Further translation for mantst.c.

	modified:   windows/graph.c

1. Added compile time dialog error setting.

2. Added variable flag for dialog error.

3. In grawrterr(), send errors to dialog or to console window
depending on dialogerr. The default is dialog.

4. Add parsing in config file for dialogerr.

Notes:

Sending runtime errors to a dialog is a better match for graphical
programs. Also, if the program is executed from a file manager or
similar, there is no assigned output (we need to test this case).

The standard Unix/Linux method is to redirect stderr to a log file.