Added mantst to makefile.
 Changes to be committed:
	modified:   Makefile

1. Added mantst to makefile.

Notes:

The formula for mantst removes a lot of dependency detail by tracing
the dependencies up through the tree. For example, services.h and
services.c don't need to be in the list because those objects are
collected into the library, and are listed as dependencies. Services.h
is a dependency of services.o, etc. config.c was actually getting
double included.

All of the formulas need to be examined and rewritten like this.
The result should be a massive simplification of the makefile.

Also, config.c should be in the libraries. It is an optional facillity,
but then most or all of PA is optional. The libraries exist as a
collection of such facillities. The user can either create a special
build to cut these down, or we can redo the make file to cut these
out with option flags.