Fixed missing pa_ coining; fixed compile errors.
 Changes to be committed:
	modified:   include/sound.h

1. Fixed missing pa_ coining definitions. Note this will require
fixups in Linux as well.

	modified:   windows/sound.c

1. Fixed various compiler errors.

Notes:

windows/sound now compiles, ready for test. This version still has
a lot of stubs in it, I am not sure when these will get filled in.
Additionally, the plug-ins features are not implemented and I will
wait and see if that is really required. In linux they were necessary,
but in Windows not so much.