Refactor genwave.c; add waveout capability to sound.c
Changes to be committed:

	modified:   linux/sound.c

1. Spelling.

	modified:   sound_programs/genwave.c

1. Increased size of buffer for Windows. See note below.

2. Refactored the sine generator using Petzold's method. The old one
had a skip problem.

	modified:   windows/sound.c

1. Removed ALSA defines (obviously we are not using ALSA).

2. Added waveout device descriptor.

3. Added waveout device table.

4. Added makpcmout() routine to allocate and initialze waveout table
entries.

5. Added openwaveout() contents. It does not do much of anything, that
is all deferred to the first wrwave() call.

6. Filled out all of the waveout parameter set functions.

7. Created the wrwave() function.

Notes:

We use the waveOut functions in Windows, which only get down to about 50ms
of latency. Thus the users have to up the size of their buffers (like
genwave). There is another, separate interface in Windows which is said
to fix the horrible latency in Windows, we may implement that later.

Windows needs double buffering to make sound work. We hide the double buffer
scheme inside wrwave() by using copyout. This takes about 6us using memcpy(),
which is a small fraction of the latency, and thus is acceptable.

I looked into implementing an algorithim that accomodates double buffering,
but it saves trival amounts of time, and in any case Windows does not allow
sharing of buffers from input to output wave devices so it would have to use
copying in any case.

The buffering algorithim allows the caller to get a buffer ahead and then
blocks. This block could be more efficient. A callback and signal system
would not waste polling time. This is a future improvement.

With this the waveout functionality is complete.