Return true parameters for wavein.
Changes to be committed:

	modified:   windows/sound.c

1. Add wave in device handle to devptr.

2. Add wave in device table.

3. Clear the wave out handle.

4. Create makpcmin() routine, to allocate input wave table entries.

5. Create fndwavfmt() routine, to determine wave properties from
windows device attribute set.

6. Create loadinpar() routine, to set input wave parameters if not
already set.

7. Fill out wave open, close, and parameter read routines using
real data.

Notes:

1. Windows (apparently) uses a subset of possible audio formats
(vis a vie Linux). The issue with windows is they add new interfaces
instead of fixing the old ones, so further research might turn them
up. For example, I can't believe in this day and age that stereo
(2 channels) is the only option.

2. Linux/Alsa didn't really resolve the issue of if input parameters
can be read from a closed wave input file. I decided that it does not
need to be open. The reason is that open is an action, and (for
example) might reserve the device to a particular program. This should
not need to happen for just reading the device characteristics.

3. I get the distinct feeling that Windows lists all possible formats
for input devices simply because it can do conversions. Read my
previous (Linux) diatribes on this nonsense. If you cannot determine
the native format of an audio stream, you will be converting it to
a different format without knowing it, and unnecessary conversion is
not good. From the Linux implementation, my correct answer is that
output devices are free to adapt to any input stream, but input
streams should tell you what their native formats are! I'll be
looking for that information.