Change wrwave() to use same buffering as rdwave; increased buffer size.
 Changes to be committed:
	modified:   windows/sound.c

1. Changed input wave buffer size INPSIZ to general WAVSIZ.

2. Increased the buffer to 46ms. This is what the study program said
was the first stable power of 2 buffer.

3. Refactored wrwave() to match the preallocated buffer and copy in
method used in rdwave. The theory is the input and output buffers
must be matched to work properly.

Notes:

This should close out rdwave() and wrwave() functionality with the
waveIn/waveOut API. Its not showing good latency, and there is another
API. Further testing will show if that is worth the effort.