Added synthesizer input capability.
 Changes to be committed:
	modified:   windows/sound.c

1. Add synthesizer input dump capability.

2. Added MIDI message queue size and buffer size.

3. Added synthesizer in timer.

4. Added MIDI input device table.

5. Added makmidinp() routine to init MIDI input devices.

6. Added dmpseq() routine, dump sequencer entries, from the Linux
version.

7. Added dmpseqlst() from same as above, but unused at the moment.

8. Replaced midtab[] with midouttab[] to make room for midinptab[].

9. Added starttimein(), stoptimein() and curtimein() routines contents.

10. Added MidiInProc() callback, which takes the MIDI short messages
from windows and places them in an input queue for the channel. Note
that we store the time stamp, but presently it is unused.

11. Filled out opensynthin().

12. Added parseq() routine. It takes the 1-3 bytes of a windows short
message and generates a sequencer entry based on that. This is a modified
routine from the Linux version.

13. Filled out the rdsynth() routine. It basically just looks at the
input message queue and gets the next entry, then encodes it in
sequencer format and returns it.

Notes:

The synth input and synth output now works and is tested. This basically
finishes the sound.c module, with the exception of odds and ends, like
better error checking and things like volume controls.

I would also like to implement proper thread locks across the interface.