Added xflush to timer.
 Changes to be committed:
	modified:   linux/graphics.c

1. Added XFlush() to timer in sig_handler().

Notes:

1. Didn't pass scntst as in last note. Combination of dropped letter and failure
to disable stupid mode. Timer passes test.

2. Next is screen select.

3. Timers work by reentering xwindows code, which is supposed to be a no-no in
both xlib and xcb (the only difference being that xcb would lock it out). This will
be a point of ongoing research in the code as if this causes issues.

Inability to implement timers is a major issue in x11. The workaround would be to
use polling implemented with xpeekevent(), which would obviously be very inefficient.
One measure would be to only use polling if their are timers active.

Note that this would also apply to joysticks, since we would have to periodically
get status on the stick using a timer.