Added reverse, underline and strikeout attributes.
 Changes to be committed:
	modified:   linux/graphics.c

1. curdrw() needed to compensate for reverse mode.

2. plcchr() implements underline and strikeout attributes. Note it will do both
at the same time! (woo hoo). I had to double up the lines to show on my high DPI
display. This may be a problem on low DPI monitors.

3. Implemented reverse() routine.

4. Implemented underline() routine.

5. Implemented strikeout() routine. Note that terminal_test has no function for
this.

6. Implemented standout() routine, which just calls reverse().

7. Removed the indisp clause from several routines. Its not necessary,
everything is copied from the bufffer. This may change when unbuffered mode is
implemented.

Notes:

terminal_test works complete now. I have seen bugs with terminal applications,
I am going to push these out for later (note the Windows implementation has
plenty of bugs as well).

The general plan is to finish linux/graphics module and head for a release.