Large refactor along Windows lines.
The code was using expose events to draw virtually everything, which was carried
over from the original xtest.c study. This is the most inefficient way to draw.
We change to drawing everything, and doing that twice, once for the buffer and
once for the screen, in case the buffer is in display.

This is how the Windows version works.

 Changes to be committed:
	modified:   linux/graphics.c

1. Removed expose event generators.

2. Redirect curdrw() to screen. Now the cursor is only on the active screen, or
does not appear at all.

3. Restore() now removes the cursor while copying.

4. iclear() now fills the screen, if active, instead of copying from the buffer.
this is more efficient and does not need a buffer to be active.

5. Refactor scrollg() to scroll the buffer and present, if it exists, or
otherwise scroll the screen directly.

6. plcchr() refactored to draw to both buffer and screen.

7. Implemented baseline().

8. Refactored wrtstr() for double draw.

9. Refactored line() for double draw.

10. Refactored rect() for double draw.

11. Refactored frect() for double draw.

12. Refactored setpixel() for double draw.

13. Implemented chrsizx() and chrsizy().

14. xwinevt() has extra flush for debugging, will be removed.

	modified:   windows/graphics.c

1. Comments.