Applied "xor" fix for characters.
 Changes to be committed:
	modified:   linux/graphics.c

1. Applied the "xor" fix for characters. One issue with applying the foreground
and backgrounds for characters seperately (necessitated by XDrawImageString()'s
lack of function ability) is that the foreground xors the background just laid
down. We fix that by drawing the foreground text over the background in the same
background colors in xor mode. after drawing the background. This restores the
background under the text (and only under the text). This needs fixup for and
and or.

Notes:

wrtstr() needs to be updated to the same procedures as plcchr().