Refactor wrtstr() with attributes/colors from plcchr().
 Changes to be committed:
	modified:   linux/dumpsynthplug.c

1. Incorrect comment.

	modified:   linux/graphics.c

1. Refactored wrtstr() to work like plcchr(), with colors and attributes.

	modified:   linux/sound.not

1. Noted CC issue. See notes.

	modified:   sound_programs/connectmidi.c

1. Added test route to dumper. Note that this only works on Linux in the present
version.

Notes:

1. wrtstr(), which started as a side note to terminal, will be used to fix the
"kerning paradox", which is that single character placements cannot (in an
obvious way) consider kerning. Note that X11 does not feature any particular
support for kerning, or justification either.

2. CC messages are currently dropped/mishandled on input. The issue is, as I
discovered on an Akai keyboard, there is no particular relationship between CC
assignments and CC functions (indeed, some of the controls on the Akai are
left blank). Thus we need a way to reassign such controls. The question is if
this is a local parameter thing or an organized part of the standard.