Finished the font selection system.
 Changes to be committed:
	modified:   linux/graphics.c

1. Turned font up to 15 on my display. See below.

2. Added "average width" field.

3. Moved font load into setfnt() from opnwin().

4. In opnwin(), moved parameter sets up so setfnt() could see them.

Notes:

I attempted to characterize the size issue at multiple resolutions. They were
generally wrong. The answer seems to be to let the user override the dimensions,
and select override values as required. Thus it was time to just tweak my
display and move on (sorry). A point size of 15 matched XTERM exactly on my
system.