Get ready for image scaling.
 Changes to be committed:
	modified:   linux/graphics.c

1. Added include for Xutil.h. XDestoryImage() used a macro there.

2. Added next pointer to pict.

3. Added picptr defintion.

4. pictbl[] is not pointers to picture entries.

5. Added frepic free picture entry list.

6. Added getpic() and putpic() routines to get an free picture entries.

7. Added delpic() to delete all active pictures (scaled) under the given logical
number.

8. Various pictbl[] references converted to pointers.

9. loadpict() uses delpic() now.

10. Refactor loadpict() to create a new picture entry and use that.

11. picture() searches the pictbl[] list for matching scaled images.

12. Cleared free pictures list.

Notes:

The idea is that for each rescaled picture we make and draw an XImage. This is
deemed more efficient since in most cases, we will draw the image at least
twice, one to buffer and one to screen. Additionally, if the client requested
a picture at scale, its possible that will get repeated. Scaling is expensive.

This is all client transparent.