Use search to find correct font size; move font metrics to config; ratio test grid.
 Changes to be committed:
	modified:   linux/graphics.c

1. Move print font metrics flag into config variable.

2. Create new setxlfd() routine to handle XLFD string creation for font. Uses
subset of setfnt() code.

3. Refactor setfnt() to search for proper font by size. See notes.

	modified:   petit_ami.cfg

1. Added font metrics flag.

	modified:   tests/graphics_test.c

1. Refactor grid() to use ratios.

Notes:

For a given pixel height, xwindows was giving characters that exceed the
bounding box, which is against PA rules. Turns out there is no way using XLFD
to exactly specify the net bounding box, you have to request a font, then use
the resulting parameters of the font to see if it fit into the bounding box.

We refactored the code to search for the font that fits the bounding box by
trial and error. It remains to see how efficient this is (I suspect not), and
thus we may have to cache font selections.