Refactor font selection algorithim; Add strikeout test.
 Changes to be committed:
	modified:   linux/graphics.c

1. Added STRIKE line placement define.

2. Added xcaplst entry structure.

3. Added caplst list of capabilities to font entries. This contains a list of
all possible combination of capabilities.

4. Added prtatset() diagnostic, prints all PA attributes in set.

5. Add prtxcset() diagnostic, prints all XWindows font capabilities in set.

6. getfonts() is refactored to generate a list of all possible capabilities.

7. Refactor selxlfd() to take a set of capabilities, and generate the font
selector string based on that.

8. Add fndxcap(), takes a PA attribute set and a XWindow capabilities set and
finds an intersection of the sets that is also mutually exclusive between modes
(like bold and xbold).

9. Add matchcap(), takes a capabilities set and finds the matching font
capabilities set from the list in each font.

10. Add fndxcapp() that takes a font entry and an attribute set, and
progressively fits the attributes to the capabilities.

11. Refactor setfnt() to fit the current attributes to what the selected font
is capable of (see also notes).

	modified:   tests/graphics_test.c

12. Formatting.

	modified:   tests/terminal_test.c

13. Added strikeout mode to font modes test.

Notes:

Because real XWindows fonts don't cover all modes and are sometimes mutually
exclusive, we needed a font fitter algorithim. An example would be a font that
has bold and extended based on the idea that the font gets wider when bold.

The algorithim used tries to fit the attributes into the font capabilities, and
if it does not work, removes attributes until it does. This is legal PA, because
the spec does not garantee any particular attribute will have effect, or that
more than one attribute can be set at a time. It's best effort.