Added MT locking to all XWindows calls.
 Changes to be committed:
	modified:   linux/graphics.c

1. Add pthreads include.

2. Add XWLOCK() and XWUNLOCK() macros. Implementation as a macro allows it to be
removed if required. However, my tests have already shown that locking is
required (despite the shift to XCD). See also notes.

3. Added ulock and unlock sequences accross all XWindows calls. Some were able
to be grouped, however, the rule is never to take this lock into a subroutine
(a rule that got broken once or twice).

	modified:   linux/network.c

1. Added pthread include. I suspect this module was getting pthreads indirectly.

Notes:

1. Fell off the testing curve, sorry. The last commit fails testing. This
underlines the idea that basic testing should always occur before commits.

2. An analysis was done on the need to lock XWindows calls. Note that there is
considerable discussion about the need to lock XWindows and how XCD addresses
locking. However, there is no garantee that XCD is in use, and in any case,
my tests found that it does not help. Xevent calls were found to be corrupted
in multithreading context.

For all X calls, there is no expectation that the call will hold waiting for
I/O or similar, although I suppose it could wait for output to the server. That
leaves XNextEvent() calls, but we only call XNextEvent() when we know that there
are input queue events (done via select()). Thus this is non-blocking.

3. There remain quite a few other locking contexts to perform. These include
tables, and locking each window/screen structure, etc.