Added virtual events to C++ terminal interface; remove remaining pa_ prefixes; added eventsover() to terminal.h.
 Changes to be committed:
	modified:   cpp/terminal.cpp

1. Added global hook variables for virtual events.

2. Added convertions for C++ types to terminal types.

3. Add virtual setup to term constructor.

4. Added virtual callbacks.

5. Added termCB event fanout function to term class.

	modified:   hpp/terminal.hpp

1. Went back to old method of parallel declarations for C++ and terminal. This
is easier deal with, only requires conversions. The drawback is that they must
be maintained in parallel as well.

2. Added eventover() and eventsover() declarations.

3. Added virtual events functions.

	modified:   include/terminal.h

Notes:

This finishes out a working implementation of C++ terminal. The remaining issues
are streams and exceptions (and exceptions are not implemented in the base C
model either).

Will feel more free to implement C++ across the whole interface when we have
proven that streams works.