Add C++ compile mode to make; Add event overrides to linux/graphics.
 Changes to be committed:
	modified:   Makefile

1. Add CPP compiler macro.

2. Added cpp/terminal.o as target.

3. Added cpp/terminal.o to all linux libraries. See notes.

4. Added targets for test++, testc++ and testg++. These are the plain, terminal
and graphical versions of test.

	modified:   linux/graphics.c

1. Add master and fanout event tables.

2. Add default vectored event handler.

3. Moved existing pa_event() to ievent(). The file parameter appears unused, but
of course that will be needed to find which window is meant.

4. Added the pa_event() handler from Windows.

5. Added pa_eventover() from Windows.

6. Added pa_eventsover() from Windows.

7. Init the event tables.

Notes:

C++ modes will be disabled by flag in the future from the make. I originally had
plans to do a separate make for C++, but the wrapper went in so smoothly and
seamlessly that it just made sense as part of the whole package.