Adjust min/max/norm test for buffering; fix unterminated strings; fix compile errors; add new win messages.
 Changes to be committed:
	modified:   tests/management_test.c

1. Redraw on changes to screen. See notes.

2. Fix unterminated string prints.

3. Fix compile errors hidden by #ifdef.

	modified:   windows/graphics.c

1. Add new definitions in prtmsgstr(). Hey, its been more than 20 years.

Notes:

The fact that Windows buffers all surfaces, as all interfaces now do, changes
the way it works a bit. The old test was counting on getting a repaint for any
window move. Now we assume we need to repaint also for changes to the surface,
like (in this case), a string that needs changing.