Fixed buffer on so that it restores the previous buffer mode correctly.
 Changes not staged for commit:
	modified:   tests/management_test.c

1. Add new win fields to store buffer parameters.

2. Added comment: the "whole" parameter on restore() is only used on WM_PAINT
operations.

3. Fix incorrect comment.

4. Fix cr to cr2 copies to use memcpy(). There's a story here, see notes.

5. Save last set buffer parameters in iniscn(). And yes, iniscn() is the only
place we create buffers.

6. Fix up isizbufg() to properly dispose of screens. See also notes.

7. ibuffer() now restores the size from the saved last buffer parameters. This
a bug where the old buffer was reused, but the new sizes of the (possibly)
resized window were used.

8. ibuffer() getrs the new screen dispose like isizbufg().

Notes:

1. The use of memcpy() is according to ANSI C for copying structures. It was a bad
translation from Pascaline, but it actually did work. There are a couple reasons
why that might be so. First, structure copy could be a later extention or a
implementation specific extension. Second, it could have been optimized to fit
in a word. In any case, its not kosher ANSI C, so we aren't using it.

2. There were several valid ways to handle turing buffered mode back on. It
could have been to fit a new buffer to whatever the unbuffer size got set to,
or it could have reset the window to the old size of buffer. The second method
has the disadvantage of changing the window geometry. I decided to use the
second because the last buffer was specifically set by the user or was the
startup default. Obviously this could be overridden.

3. The buffer recycle section needs to be completed.

4. The whole mode for restore() and the WM_PAINT mode are different because in
the WM_PAINT mode Windows will break down the right and bottom edge paints into
two separate operations, but the whole mode will not. This needs to be looked
at.