Change the way resize tests work.
 Changes to be committed:
	modified:   tests/management_test.c

1. Added rewrites on resize (etresize) event.

Notes:

The theory of PA was that the screen only needed repaint in unbuffered mode if a
redraw event was sent. If the resize event actually changes the data onscreen,
as in the case of these tests, the underlying API cannot know that, so does not
issue a redraw event. The reason this worked on the older systems is that
Windows always issued redraws on resizes, even downward ones (that do not
uncover new area). Windows no longer does that because it buffers all windows,
and it knows that a window shrink does not need a redraw.

Thus I consider this to be a fault in the test.