Added test results; fix translation error; change notes.
 Changes to be committed:
	modified:   tests/graphics_test.c

1. Add test results for non-scaled mode Windows.

	modified:   tests/management_test.c

1. Fix translation error.

	modified:   tests/terminal_test.c

1. Add test results for non-scaled mode Windows.

	modified:   tests/widget_test.c

1.
	modified:   windows/graphics.not

1. Added notes and removed notes that are no longer a problem.

Notes:

1. non-scaled mode on Windows works complete but there are some testing issues
that arose.

2. I choose the scaling mode that turns off all scaling under all circumstances,
even drag to other monitor with different scaling (laptop mode). Windows
provides DPI change messages. I think the way to handle this is to issue resize
events (by user option) so that when a window is dragged to another display,
it automatically fixes the window to occupy the same DPI (note net size on
display! Not pixels!).

Turning all scaling off is the right decision, IMHO. Scaling is damage. It blurs
and distorts the figures on screen and generally reduces the quality. Its
totally unnecessary. Applications are perfectly capable of scaling themselves,
and they do a much better job of it than the OS.