Turn on scaling; add final prints for dropdown edit boxes; Add full geometry diagnostic; changes to notes.
 Changes to be committed:
	modified:   petit_ami.cfg

1. Turned scaling back on. Non-scaled mode does not work.

	modified:   tests/management_test.c

1. diagnostics.

	modified:   tests/widget_test.c

1. Added print of final value for dropdown edit boxes. See notes.

	modified:   windows/graphics.c

1. Added a full dump of the window geometry from Linux version.

	modified:   windows/graphics.not

1. Reedit and add notes.

Notes:

1. It does not look good for non-scaled mode. With scaling off, the basic
Windows functions like GetWindowRect() malfunction. I read articles that say it
is due to "invisible borders", but that does not explain what I am seeing.
The new "correct" functions use calls that the current mingw implementation
does not have. I am going to back off this idea and kick it into the future.

2. dropdown edit boxes don't meet the PA spec, and aren't consistent with other
such controls, either. I am exploring fixes with subclassing the controls.
In the meantime, the workaround is to get the contents on completion.