Add news comment; add flush to terminal diag; fixes for transparent mode; enforce global set policy.
 Changes to be committed:
	modified:   NEWS

1. Add comment to news: Linux does not support transparent mode. See notes.

	modified:   windows/terminal.c

1. Enforce global set policy by setting auto, cursor visible, attibutes and
colors globally at the same time we set screen local versions. See notes.

2. Fix geometry for transparent mode. See notes.

Notes:

1. Transparent mode is an old (like 17 years old) observation that in Windows,
at least, terminal mode can be invoked, and if you don't do anything with
terminal mode calls, the display is completely unaffected. You would not even
notice a terminal mode program is invoked. It was used by exactly one program,
pexec, a shell, to run its command line editor.

So who cares right? Well, its kinda a quality issue. If terminal mode is
completely downward compatabile, it should also be transparent. Its not in
the xterm compable terminal mode because xterm basically forces you to switch
buffers, and because (unlike windows) it does not manage buffers for you.

However, I have identified a possible method to support it, and that may
go forward on Linux. The next part of that story is that since 17 years ago,
Windows broke transparent mode, and I am trying to get it back.

2. After several weeks of examining the screen attributes set policy, I decided
that the defacto policy (what got put into the code without that much thought)
was actually not bad, and was a good way to go.

The policy(tm) was that attbutes, auto mode, colors attributes, etc, were per
screen, but set globally anytime they are changed. What this does is allow each
screen to have its own context such that if you switch to screen N, you pick up
the last set characteristics of that screen, but that if you set a
characteristic, not only the current screen gets that, but it gets set globally
as well. What does that matter if you get changed automatically per screen set?
Well, it solves the "new screen ambiguity", that is, if you create a new
screen/buffer, what are the default characteristics? Instead of getting the
power-on defaults, or an arbitrary screen, you get the characteristics you
last set.