Fix spellings; Correct screen offsets.
 Changes to be committed:
	modified:   sound_programs/play.c

1. Fixed spellings.

	modified:   windows/terminal.c

1. Corrected Y coordinate given to windows by buffer display offset. See notes.

Notes:

This is the "fixed display mode". It works fine, but it will basically run the
screen protocol in place, wherever the display window is in the buffer.
The next step is to just let windows handle scroll down operations (simple one
line down scrolls). This basically allows the terminal module to be fully
transparent.

Previously I stated that only pexec (which by the way is not in the PA
collection) was the only program to benifit from transparent mode. That was not
true. Play is a good example of a program that uses transparent mode. It uses
the timer and event model of terminal, but does not use any of the advanced I/O!
Thus it is a regular console program that just happens to use the timer.