Update test report; Double size of wave input buffer; update sound notes.
 Changes to be committed:
	modified:   TESTREPORT

1. Update test report.

	modified:   windows/sound.c

1. Doubled the size of the wave input buffer, which makes it 92ms, which is
really unacceptable, but fixes the problems with it. See notes.

	modified:   windows/sound.not

1. Added sound notes about how bad the MS wave interface is.

Notes:

1. The old input buffer size was set to the low edge of what works, 46ms.
Apparently that no longer works, so I doubled it to 96ms.

The wave interface latency is bad, and MS documentation covers this. They
provide a "low latency" interface, but of course that is incompatible with the
old API, meaning we have to recode it. Putting this off for later.

2. This commit finishes off the vs. 0.1 Windows responsibilities. Any remaining
items are pushed into the next release frame.